requirements.txt
requests~=2.28.2

pyyaml~=6.0
